Replace all devices us_ticker files with a common 16b and 32b file
