+++ targets/TARGET_STM/stm_us_ticker_32b.c
@@ -28,15 +28,15 @@
 #include <stddef.h>
 #include "us_ticker_api.h"
 #include "PeripheralNames.h"
+#include "hal_tick.h"
 
-// 32-bit timer selection
-#define TIM_MST      TIM2
+// Default is a 32bit timer
+#if !TIM_MST_16BIT
 
 static TIM_HandleTypeDef TimMasterHandle;
 static int us_ticker_inited = 0;
 
-void us_ticker_init(void)
-{
+void us_ticker_init(void) {
     if (us_ticker_inited) return;
     us_ticker_inited = 1;
 
@@ -45,26 +45,29 @@ void us_ticker_init(void)
     HAL_InitTick(0); // The passed value is not used
 }
 
-uint32_t us_ticker_read()
-{
+uint32_t us_ticker_read() {
     if (!us_ticker_inited) us_ticker_init();
     return TIM_MST->CNT;
 }
 
-void us_ticker_set_interrupt(timestamp_t timestamp)
-{
+void us_ticker_set_interrupt(timestamp_t timestamp) {
     // Set new output compare value
+// TODO: Check if still true
+#if defined(TARGET_L4)
+    __HAL_TIM_SET_COMPARE(&TimMasterHandle, TIM_CHANNEL_1, (uint32_t)timestamp);
+#else
     __HAL_TIM_SetCompare(&TimMasterHandle, TIM_CHANNEL_1, (uint32_t)timestamp);
+#endif
     // Enable IT
     __HAL_TIM_ENABLE_IT(&TimMasterHandle, TIM_IT_CC1);
 }
 
-void us_ticker_disable_interrupt(void)
-{
+void us_ticker_disable_interrupt(void) {
     __HAL_TIM_DISABLE_IT(&TimMasterHandle, TIM_IT_CC1);
 }
 
-void us_ticker_clear_interrupt(void)
-{
+void us_ticker_clear_interrupt(void) {
     __HAL_TIM_CLEAR_IT(&TimMasterHandle, TIM_IT_CC1);
 }
+
+#endif // !TIM_MST_16BIT
